+++ MordoEngine/src/Terrain/HeightMapTerrain.cpp
@@ -1,6 +1,9 @@
 #include "HeightMapTerrain.h"
+#include <iostream>
 
 HeightMapTerrain::HeightMapTerrain(const std::string& filepath)
 {
-	LoadHeightMap(filepath);
+	if (!LoadHeightMap(filepath)){
+		std::cerr << "Error: Failed to load heightmap" << filepath << std::endl;
+	}
 }